refactor: updates views folders
