+++ FuelPump/FuelPump/Resources/es.lproj/Localizable.strings
@@ -13,6 +13,8 @@
 "onboarding.selector.title" = "Elige tu tipo de combustible";
 "onboarding.selector.subtitle" = "Seleccione una opción:";
 
+"list.title" = "Encuentra las gasolineras más rentables";
+
 "gas.station.details.address" = "Dirección";
 "gas.station.details.schedule" = "Horario";
